Update TOD
I tried to mark all the changes I made, but there is a big change that I missed some of the changes since I didn't think of doing it until I finished editing... oops...

In the attribute, can you verify the calculation for Initial_Bus/Rail_FAR_Bonus? I'm not sure this is what you described but we can change it easily.

The forms look quite similar - any variations possible?

We got some long, linear parcels, might need to readjust the setbacks? or not?